+++ apps/web/src/app/components/input/input.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+import { FormControl, Validators } from '@angular/forms';
+
+@Component({
+  selector: 'app-input',
+  templateUrl: './input.component.html',
+  styleUrls: ['./input.component.scss']
+})
+export class InputComponent implements OnInit {
+  errorControl = new FormControl(null, Validators.required);
+
+  constructor() {}
+
+  ngOnInit() {}
+}