+++ libs/xui/src/input/index.ts
@@ -1,0 +1,3 @@
+export * from './input.module';
+export * from './input.component';
+export * from './input-group.component';