+++ libs/xui/src/input/input-group.service.ts
@@ -1,0 +1,6 @@
+import { Injectable } from '@angular/core';
+
+@Injectable()
+export class InputGroupService {
+  size: 'normal' | 'small' = 'normal';
+}