+++ libs/xui/src/input/input.component.ts
@@ -1,0 +1,93 @@
+import { ChangeDetectorRef, Component, Input, Optional, Self, ViewEncapsulation } from '@angular/core';
+import { ControlValueAccessor, NgControl } from '@angular/forms';
+import { TranslateService } from '@ngx-translate/core';
+import { InputGroupService } from './input-group.service';
+
+@Component({
+  selector: 'xui-input',
+  exportAs: 'xuiInput',
+  encapsulation: ViewEncapsulation.None,
+  // changeDetection: ChangeDetectionStrategy.OnPush,
+  templateUrl: 'input.component.html'
+})
+export class XuiInputComponent implements ControlValueAccessor {
+  _value?: string;
+  touched = false;
+  onChange = (source?: string) => {};
+  onTouched = () => {};
+
+  @Input() placeholder?: string;
+  @Input() disabled: boolean = false;
+  @Input() color: 'light' | 'dark' = 'light';
+  @Input() xSize: 'normal' | 'small' = 'normal';
+  @Input() dataList?: string[] | null;
+  @Input() type = 'text';
+
+  @Input()
+  get value() {
+    return this._value;
+  }
+
+  set value(v) {
+    if (this._value !== v) {
+      this._value = v;
+      this.onChange(v);
+    }
+  }
+
+  get styles() {
+    return `input xui-input-${this.color} xui-input-${this.groupService?.size ?? this.xSize}`;
+  }
+
+  get errorMessage() {
+    const msg = this.control?.getError('message');
+    return this.translation.instant(msg);
+  }
+
+  constructor(
+    private changeDetectorRef: ChangeDetectorRef,
+    private translation: TranslateService,
+    @Optional() private groupService: InputGroupService,
+    @Self() @Optional() public control?: NgControl
+  ) {
+    if (this.control) {
+      this.control.valueAccessor = this;
+    }
+  }
+
+  ngOnInit() {
+    this.control?.statusChanges!.subscribe(() => this.changeDetectorRef.markForCheck());
+  }
+
+  get invalid(): boolean {
+    return !!this.control?.invalid;
+  }
+
+  get showError(): boolean {
+    if (!this.control) {
+      return false;
+    }
+
+    const { dirty, touched } = this.control;
+    return this.invalid ? dirty! || touched! : false;
+  }
+
+  writeValue(source: string) {
+    this.value = source;
+  }
+
+  registerOnChange(onChange: any) {
+    this.onChange = onChange;
+  }
+
+  registerOnTouched(onTouched: any) {
+    this.onTouched = onTouched;
+  }
+
+  markAsTouched() {
+    if (!this.touched) {
+      this.onTouched();
+      this.touched = true;
+    }
+  }
+}